Made a template screen for the classes stuff
Sampath, if you read this, you need to do two things. First I need you to get the whiteboard thing working and maybe also the video thing. Second, is that I have the pages made and the links can just be copie from the wix site. However the pdf files need to be set up in this, rather than as the wix link. So if you do that, we are done